Don't allow attrib count to be odd
If the final count is odd, that means the user accidentally added or
removed an attribute/value, which means the whole array could be
shifted.  So if we encounter that, instead of assuming it happened on
the last cell, exit just to be safe.  This is consistent of what we do
for chances.