Use EmitSoundToAllExcept where applicable
In addition, remove FF2FLAG_TALKING as it is *completely* useless.  It's
executed synchronously so for all intents and purposes it'll always be
"false".  For it to actually have a use it would need to be unset in a
timer, but then you would need to know how long each sound plays, which
leads us to the wonderful GetSoundDuration function with the description
of "Does not work, may crash".  So nope, not going through that effort.